Fixing CodeQL. (#273)
# Describe Request

Actions workflow file not found
The Action workflow file .github/workflows/codeql-analysis.yml no longer
exists.

To stop seeing this error and alerts that were detected by this
configuration, you can delete the configuration from the menu in the
upper right of this page.

# Change Type

CodeQL config.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Chores**
* Introduced automated CodeQL analysis to continuously scan for security
and code quality issues in Go code on the main branch and on a weekly
schedule.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->