fix the calculation of the WMA
This fix #274, WMA calculation was wrong, correct formular is, i.e. WMA(5):

(P1 * 5) + (P2 * 4) + (P3 * 3) + (P4 * 2) + (P5 * 1) / (5 + 4 + 3 + 2 + 1).

This makes it perfect for the divisor to be calulated only one (gauss formula: (n * (n + 1) / 2).

The testdata has been new caclulated manually via Excel, a WMA(3) and a WMA(5) … to be safe.

Signed-off-by: Heiko Weber <heiko@wecos.de>